feat(script-py): script for multi-process ib_*_* running and result parsing
Debug for a long time. ssh has a rule: it limits the unauthorized connection by (by default) 10.
So, the ssh connection keeps disconnecting. It, in turn, causes the server to hang forever.
I though it was the bug from subprocess, but it turns out to be the problem from ssh connection.
It really takes me a lot of time handling them.

BTW, the expr results:
9-12 threads reach the maximum send/recv ops.
size -> larger, ops -> slower.
turn on the "inline" helps A LOT.