done: correct rmsg
I though rmsg has bug, and debug it for, like, 1+ days. The reason is that bench_rmsg.cpp always fails by exceeding retry count.

After a lot of debuging, I realize that my code is correct.
The sender is too fast. It will not wait until the receiver posts enough buffer. So the post_send will fail.
The possbile fix is that, the server ack clients that it is ready, before client can continue sending messages.
I will tune the parameter for an optimized performance