refine(rmsg): carefully set QP depth
the max_{send|recv}_wr capability field of QP, and
the depth of CQ,
is carefully set, so that no dram is wasted.
It slightly speeds up the message, but I am not testing it carefully.

Tomorrow, I will bench the rmsg.
On weekend, I will begin the coding of Pratronus.