feat(latency): finish latency monitor of bench_patronus_basic
This commit takes me hours.
Do not open timer for EVERY coroutine, this will slow down the program a lot. Because of the scheduling of coroutine,
the slowdown of peer coroutine also affects the latency of the current coroutine.

In this commit, we determine the enabling of timer by `bool enable_trace = xxx`.
Only enable the timer when `enable_trace` is true. Therefore, the reported results will be accurate.