prepare(header), fix(RemoteConnection): prepare for the swap area of Lease
preapre(header):
now the management of DSM buffer is perfect, and every abstraction does it work.
User provide a PatronusConfig to Patronus, assigning how much buffer it wants.
Patronus pass a DSMConfig to DSM, assigning how much buffer and reserved buffer it wants.
In essense, the DSM buffer is three part.
[DSM reserved][User reserved][buffer]
- DSM reserved: reserved for bootstrap exchange data. used by DSM
- User reserved: reserved for DSM user, i.e., Patronus. Patronus uses it as the header pool
- buffer: the actual buffer. DSM pass to Patronus, Patronus passes to the end user.
The config file is perfect to assign what each layer wants

fix(RemoteConnection):
ibv_ah is bind to PD. However, when we freeing PD without freeing the AH,
the NIC does not complain about it (It should! resource busy!)
If PD is freed, the later try to free AH will crash. And it is very difficult to debug!
If we reinit_dir, the DirectoryConnection will reset, the PD will free.
We have to manually free the AH before freeing the PD (when free the context).
That's how I fix this bug. Free the AH, and set the ah = nullptr.