temp(one-sided extend): WIP client extending lease
Client uses one-sided CAS to extend the lease.
What client does is to CAS a field to make it double.
If success, the extending is deemed successful. The server should respect it.
If failed, the extending is failed. The client could also use the lease till the end.

This commit passes the correctness of buffer_cas.
Will test whether lease extends works well (protection_region cas).