feat: race hashing support multi-threading
pass correct_race_op, which is a multi-thread test.

Change the API:
introduce Slot, SlotView and SlotWithView
Slot is a write-able object representing the slots.
SlotView is a read-only view of the slot.
SlotWithView a pair with Slot and SlotView

When client RDMA read the combined buckets, it searches for any Slot
candidates it wants. It is represented by SlotWithView.
Then the client tries to modify the slot by calling
slot_with_view->modify, which issues RDMA CAS to try modifying the
remote fields.