plz revert me: migrate bug
If we do like Race hashing paper:
1. insert dst subtable into directory
2. do the copy from src to dst
3. delete items from src

There will be rc:
The expansion client is copying data to dst
The reader updates its cached directory, and issue the read to the new
dst subtable. Them it loses its expected objects.

So the modifications:
Finish the copying before inserting dst into the directory