fix: ngx_allocator, tagged_ptr, ah
ngx_allocator:
memory leak, because not resetting the shared_ptr in
p->internal_allocator.

tagged_ptr:
memory leak because not freeing

ah:
when directory is dctor, the ah from ALL NODEs needs to be dctor,
because they all depends on the context from the directory.
So add a loop for node_id from 0 to machine_nr - 1. forgot that loop