feat: using backup QP for fast recovery
Patronus & DSM:
Add abstraction of ThreadDesc, which is the the thread_local state for
running.
We can pre-alloc the desc and store in the queue. When we need fast
recovery, we can switch to the ready QP by overwriting the current
thread desc by the backup thread desc.

The register-thread procedure of Patronus & DSM is still hacky, because
a) we need to keep register thread idempotent (we have to register
thread in the internal stuffs, and can not unregister the thread)
b) we need to distinguish the idempotent registering threads, and the
none-idempotent preparing thread desc.

The still have lots of problems by hacky impl:
The resources (i.e. buffers) can not be gc-ed under the new descs.
The tid will change, just take cares.

Renaming:
the thread_locals now rename from `name` to `name_`. should be
`name_tl_` better, but won't fix them now.