trial and numbers
API(patronus): AcquireRequestFlag::kDebugServerDoNothing
server will do nothing.
like no-bind-any, but it further bypass allocating lease_ctx
like alloc-only, but it will not alloc.

When we totally skip the RPC of relinquish, from bench_patronus_basic,
the `w/o(*) no-rel`, we got 10 Mops.
However, we normally can not get this, because when u skip relinquish,
resource will use up eventually.
Alternatively, we can use auto-expire to reduce the network overhead.
However, it brings overhead of DDL Manager, which is a heap and takes
O(NlogN). gprof reports that it takes a lot of time.

Try this: let a specific coroutine to do the DDLManager poll.