prepare for refactor of RH
a) enable RO, WO and R/W. Before that, we need to solve the problem of
out-of-memory. We can implement a special allocator which recyle its
memory. To support this, we should also modify RH's `matching kvblock
algorithm` so that it ignore the corrupted kv block it encounters

b) then, get the basical results of RO, WO and R/W for
bench_race_mortgate and bench_race_rdma. See which is better

c) Try to implement batch for insert/allocation. Implement single huge
RO handle for get operation. This may boost performance, making Patronus
and unprotected closer.