tmp: pass compiling
Huge refactor to serverless, because it is so tedious for the serverless
functions to use Patronus to manually do the IO.
Now, we abstract out the IO into the Parameters object. Through this
object, the lambda can only focus on the execution logic, rather than
the actual manipulation with Patronus.

Should rewrite and pass correct_launcher_*, then continue to finish the
image lambda benchmark.

How serverless uses launcher may possibly be wrong. review it.