perfect MW bench for bench_prot_lat
bench_prot_lat.cpp:
Now the bench for MW is perfect. It is multi-threading, and the
situation should be very real.

We find that the use of MW have locality issue:
If we reuse MW among a small number of pool, it has better performance.
On the other hand, round-robin within a large pool is poor.